Only show widget if not running in mirror (#54)
On template creation, the ‘velocity’ template contents are no longer
added to the DOM.  The contents are only added if it is subsequently
determined that the browser session is NOT running in a velocity
mirror.  Users may see a very short delay before the status widget
appears, depending on how long it takes for the call to the
`velocity/isMirror’ takes to return a value.

This logic ensures that browser sessions running in a mirror will NEVER
display the status widget.  This is very helpful for automated
end-to-end UI tests, because it ensures the widget can never obscure
other elements underneath it.